requirements.txt
streamlit
pymongo
requests
matplotlib
plotly
python-dotenv